add solr search

+++ src/main/java/com/services/CommentService.java
@@ -1,6 +1,7 @@
 package com.services;
 
 import com.DAO.CommentsRepository;
+import com.exceptions.NodeException;
 import com.models.Comment;
 import com.mongodb.client.gridfs.model.GridFSFile;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,13 +13,15 @@ import java.util.List;
 
 @Service
 public class CommentService {
+    private final static String WRONG_NODE_ID_MESSAGE = "no such node exists";
+
     @Autowired
     private NodeService nodeService;
 
     @Autowired
     private CommentsRepository commentsRepository;
 
-    public void addComment(Comment comment, String nodeId) {
+    public void addComment(Comment comment, String nodeId) throws NodeException {
         GridFSFile mainVersionParentNodeId = nodeService.getMainVersionParentNodeId(nodeId);
         comment.setNodeId(mainVersionParentNodeId.getId().asObjectId().getValue());
         commentsRepository.save(comment);
@@ -28,19 +31,22 @@ public class CommentService {
         commentsRepository.deleteById(commentId);
     }
 
-    public List<Comment> getAllComments(String nodeId) {
+    public List<Comment> getAllComments(String nodeId) throws NodeException {
         GridFSFile mainVersionParentNodeId = nodeService.getMainVersionParentNodeId(nodeId);
         return commentsRepository.findByNodeId(mainVersionParentNodeId.getId().asObjectId().getValue());
     }
 
-    public List<Comment> getAllCommentsPage(String nodeId, int page, int size) {
+    public List<Comment> getAllCommentsPage(String nodeId, int page, int size) throws NodeException {
         GridFSFile mainVersionParentNodeId = nodeService.getMainVersionParentNodeId(nodeId);
         Page<Comment> comments = commentsRepository.findByNodeId(mainVersionParentNodeId.getId().asObjectId().getValue(), PageRequest.of(page, size));
         return comments.getContent();
     }
 
-    public void deleteAllComments(String nodeId) {
+    public void deleteAllComments(String nodeId) throws NodeException {
         GridFSFile mainVersionParentNodeId = nodeService.getMainVersionParentNodeId(nodeId);
+        if (mainVersionParentNodeId == null) {
+            throw new NodeException(WRONG_NODE_ID_MESSAGE);
+        }
         commentsRepository.deleteAllByNodeId(mainVersionParentNodeId.getId().asObjectId().getValue());
 
     }
